requirements.txt
pandas==1.3.5
numpy==1.21.5
matplotlib==3.4.3
seaborn==0.11.2
pyarrow==15.0.1

# Snowflake Dependencies
snowflake-sqlalchemy==1.5.1